code after my own review
