+++ src/components/inputRef.tsx
@@ -1,0 +1,5 @@
+import { useRef } from 'react';
+
+export function useInputRef() {
+  return useRef<HTMLInputElement>(null);
+}
